compile as a single translation unit
No more header nonsense or Makefile garbage. Look at the diffstat
on this one - even a tiny program like this had >100 lines of
useless boilerplate. I also have better ways of separating the
platform layers nowadays but we will see if I care enough to
change it.

The build.sh script is added for documentation. Its also not
really important.